+++ My-App/src/app/app.module.ts
@@ -3,63 +3,25 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './header/header.component';
-import { RecipesComponent } from './recipes/recipes.component';
-import { RecipeDetailComponent } from './recipes/recipe-detail/recipe-detail.component';
-import { RecipeListComponent } from './recipes/recipe-list/recipe-list.component';
-import { RecipeItemComponent } from './recipes/recipe-list/recipe-item/recipe-item.component';
-import { ShoppingListComponent } from './shopping-list/shopping-list.component';
-import { ShoppingEditComponent } from './shopping-list/shopping-edit/shopping-edit.component';
-import { DropDown } from './shared/dropdown.directive';
-import { ShoppingListService } from './shopping-list/shopping-list.service';
-import { RecipeStartComponent } from './recipes/recipe-start/recipe-start.component';
-import { RecipeEditComponent } from './recipes/recipe-edit/recipe-edit.component';
-import { AuthGuard } from './auth-guard.service';
-import { AuthService } from './auth.service';
-import { canDeactivateGuard } from './can-deactivate-guard.service';
-// import { ResolveGuard } from './resolve-guard.service';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { RecipeService } from './recipes/recipe.service';
-import { DataStorageService } from './shared/data-storage.service';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { AuthComponent } from './authentication/authentication.component';
-import { AuthInterceptorService } from './authentication/auth-interceptor.service';
+import { DropDownDirective } from './shared/dropdown.directive';
+import { HttpClientModule } from '@angular/common/http';
+import { RecipesModule } from './recipes/recipes.module';
+import { ShoppingListModule } from './shopping-list/shopping-list.module';
+import { AuthModule } from './authentication/auth.module';
+import { CoreModule } from './core.module';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    HeaderComponent,
-    RecipesComponent,
-    RecipeDetailComponent,
-    RecipeListComponent,
-    RecipeItemComponent,
-    ShoppingListComponent,
-    ShoppingEditComponent,
-    DropDown,
-    RecipeStartComponent,
-    RecipeEditComponent,
-    AuthComponent,
-  ],
+  declarations: [AppComponent, HeaderComponent, DropDownDirective],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    FormsModule,
-    ReactiveFormsModule,
     HttpClientModule,
+    ShoppingListModule,
+    RecipesModule,
+    AuthModule,
+    CoreModule,
   ],
-  providers: [
-    ShoppingListService,
-    AuthGuard,
-    AuthService,
-    canDeactivateGuard,
-    // ResolveGuard,
-    RecipeService,
-    DataStorageService,
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: AuthInterceptorService,
-      multi: true,
-    },
-  ],
+  providers: [],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
