+++ My-App/src/app/authentication/auth.module.ts
@@ -1,0 +1,20 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { RouterModule } from '@angular/router';
+import { AuthComponent } from './authentication.component';
+
+@NgModule({
+  declarations: [AuthComponent],
+  imports: [
+    FormsModule,
+    CommonModule,
+    RouterModule.forChild([
+      {
+        path: 'auth',
+        component: AuthComponent,
+      },
+    ]),
+  ],
+})
+export class AuthModule {}
