+++ My-App/src/app/shared/dropdown.directive.ts
@@ -3,7 +3,7 @@ import { Directive, HostListener, HostBinding } from '@angular/core';
 @Directive({
   selector: '[myDropDown]',
 })
-export class DropDown {
+export class DropDownDirective {
   @HostBinding('class.open') isOpen: boolean = false;
   @HostListener('click') toggleOpen() {
     this.isOpen = !this.isOpen;
